Cleaned up the ClientCommunicator "register to write" gear to make it MUCH clearer; the only reason I had the "ClientEvent" class was to manage both Client and Server interest-op changes: since we're only interested in Server "socket ready to write" events we can simply pass in the socket.
Also started working on having multiple, sequenced ENTRY commands for a context, as well as parameters for context commands; I need a way to get from the first context to the login context and I'm planning to use a 2nd ENTRY command called "SetContext" with a parameter for the ID I want. 

Removed the .iws file as I think it's giving me fits.